refactor: break the context ↔ element circular import via a leaf transform module
The runtime value cycle crossed context ↔ element on two edges:
context.ts imported applyElementTransform from core/element, and
element-transform.ts imported resolveRotation/resolveTransformOrigin from
the ../context barrel — which dragged in the whole Context class.

Both resolvers are pure (they depend only on degreesToRadians and
typeIsNumber), so they move to a new leaf module context/transform.ts that
imports neither context.ts nor element*. context.ts re-exports them and the
context barrel picks them up directly, so the public surface is unchanged;
element-transform.ts now imports the leaf instead of the barrel, cutting the
cycle. Pure move — no behaviour change.
