+++ packages/core/src/context/index.ts
@@ -1,4 +1,5 @@
 export * from './path';
 export * from './text';
+export * from './transform';
 export * from './context';
 export * from './types';
