+++ packages/core/src/core/element-transform.ts
@@ -7,7 +7,7 @@
 import {
     resolveRotation,
     resolveTransformOrigin,
-} from '../context';
+} from '../context/transform';
 
 import type {
     Matrix,
